Add new state related module
